Float scores, scraper fix, stock-priority sorting
- Score field changed from int to float (handles 83.5, 85.75 etc.)
- Scraper parses score as float, edit input step=0.25
- Sidebar sort always groups by stock priority (at home > in store >
  archive), selected sort key only breaks ties within each group
- Tested all 18 MadHeads catalog products — 0 issues

+++ backend/app/models.py
@@ -63,7 +63,7 @@ class Coffee(Base):
     roast_level = Column(String, nullable=True)
     roastery_url = Column(String, nullable=True)
     image_url = Column(String, nullable=True)
-    score = Column(Integer, nullable=True)
+    score = Column(Float, nullable=True)
     sweetness = Column(Integer, nullable=True)
     acidity = Column(Integer, nullable=True)
     bitterness = Column(Integer, nullable=True)
